implemented Project Json
Implemented projects.json, as well as modularity to automatically create projects sections as the data is added to the .json. (Potentially create a scrapper to generate the json off of an excel sheet for easier project adding?). Also implemented single project functionality, so that specific projects can be chosen allowing the creation of the featured projects section. Increase robustness of javascipt json project scrapper for projects that dont include a description. Maybe some sort of "in the work" or "TBD" functionality